Adding support for multiple repos; adding config file for repo specific params (#26)
The server is now more dynamic and configurable. It also has other improvements that I will state below:

Config.plist has been created for server specific configurations. If it's the GitHub app ID, cloud storage bucket, PEM file location, and other configurable settings.
There is now a ConfigManager singleton that fetches the configuration from the plist, and there is also a DefaultConfigParams struct that has default params to fall back on if none are set for specific things.
There is now a GithubManager singleton that based on the installationID received from the webhook, creates a github instance that is specific for that repo (installation). So if 2 repos have the Github App installed, it creates 2 instances for the GitHub API, because each repo has a different configuration and a different access_token.
I removed the recursive logic when a request fails due to being unauthorized. The reason being is that it may in some weird cases go into an infinite loop if the token is refreshed but still it returns as unauthorized. Instead I created a block that builds each request and is invoked initially and once more after refreshing the token.
I removed the dependency of any need to set a project path, it automatically finds the path by invoking the swift #file invocation. (this is done in ConfigManager.getProjectDirectory).
repository_url has been added to the IssueData class as it is needed for some API calls, and removes the dependency of explicitly telling the server your repo name.
closes #17 #18