Iterations to the Config API. (#42)
This change removed all singleton behavior from the Config class, in favor of making it work as a well-formed "holder of data" that must be fully successfully initialized and passed around to other objects. This reduces cross-module global dependencies and allows us to better define the public API as a set of typed non-optionals. In the future, if we re-enable plist loading behavior, we can have a failable initializer that attempts to set all of the required config parameters from a plist.

The config API was previously loading some data from Config.plist while defining other defaults in source. Both files were checked in to the repo, meaning there was a layer of indirection that was increasing the complexity of the code throughout the repo without actively providing any benefits. This resulted in code like the following:

```swift
guard let githubAppIDStr =  ConfigManager.shared?.configDict["GITHUB_APP_ID"] as? String
```

Which has now been replaced with code like the following:

```swift
config.githubAppId
```

The intent of the config file is to allow someone else to deploy the GitHub app to their own server with their own custom configuration. It is unlikely that we will support this particular model in the immediate future. Once we do, though, we can restore the plist loading and indirection.