Extract issue labels from complex structures. (#44)
The previous code was assuming that the labels array was an array of strings. I don't know if this is ever the case; generally-speaking, an issue's labels key contains an array of dictionaries.

Prior to this change, issue labels were not being extracted from issue re-labeling events.

After this change, an issue will extract the issue labels from the issue object.